+++ framework/CHANGELOG.md
@@ -12,6 +12,7 @@ Yii Framework 2 Change Log
 - Enh #20878: Add `@param-out` annotation for `$models` in `yii\db\BaseActiveRecord::loadRelationsFor()` (mspirkov)
 - Bug #20878: Fix `@var` annotation for `yii\db\Query::$from` (mspirkov)
 - Enh #7616: Add `yii\web\ErrorHandler::EVENT_AFTER_RENDER` and `yii\web\ErrorHandlerRenderEvent` to post-process rendered HTML error output (terabytesoftw)
+- Bug #20885: Fix invalid jQuery selector built from empty `yii\widgets\ActiveForm::$errorSummaryCssClass` and `yii\widgets\ActiveField::$errorOptions['class']` (WarLikeLaux)
 
 
 2.0.55 May 09, 2026